+++ src/pages/_app.tsx
@@ -1,0 +1,13 @@
+import "../../styles/globals.css";
+import type { AppProps } from "next/app";
+import { RecoilRoot } from "recoil";
+
+function MyApp({ Component, pageProps }: AppProps) {
+  return (
+    <RecoilRoot>
+      <Component {...pageProps} />
+    </RecoilRoot>
+  );
+}
+
+export default MyApp;
