+++ src/pages/index.tsx
@@ -1,0 +1,59 @@
+import type { NextPage } from "next";
+import Head from "next/head";
+import Image from "next/image";
+import styles from "../../styles/Home.module.css";
+import { Map } from "../components/map";
+
+export const getStaticProps = async () => {
+  const markers = [
+    {
+      id: 1,
+      name: "東京タワー",
+      lat: 35.658,
+      lng: 139.74,
+    },
+  ];
+
+  const center = {
+    lat: 35.658584,
+    lng: 139.7454316,
+  };
+
+  return {
+    props: {
+      markers,
+      center,
+    },
+  };
+};
+
+const Home: NextPage = ({ markers, center }: any) => {
+  return (
+    <div className={styles.container}>
+      <Head>
+        <title>Create Next App</title>
+        <meta name="description" content="Generated by create next app" />
+        <link rel="icon" href="/favicon.ico" />
+      </Head>
+
+      <div>
+        <Map zoom={9} markers={markers} center={center} />
+      </div>
+
+      <footer className={styles.footer}>
+        <a
+          href="https://vercel.com?utm_source=create-next-app&utm_medium=default-template&utm_campaign=create-next-app"
+          target="_blank"
+          rel="noopener noreferrer"
+        >
+          Powered by{" "}
+          <span className={styles.logo}>
+            <Image src="/vercel.svg" alt="Vercel Logo" width={72} height={16} />
+          </span>
+        </a>
+      </footer>
+    </div>
+  );
+};
+
+export default Home;
